+++ source/root/usr/lib/logdata-anomaly-miner/aminer/YamlConfig.py
@@ -464,7 +464,7 @@ def build_analysis_components(analysis_context, anomaly_event_handlers, atom_fil
                                     set_lower_limit=item['set_lower_limit'], set_upper_limit=item['set_upper_limit'],
                                     learn_mode=learn, output_logline=item['output_logline'], ignore_list=item['ignore_list'],
                                     constraint_list=item['constraint_list'])
-            elif item['type'].name == 'EventCountDetector':
+            elif item['type'].name == 'EventCountClusterDetector':
                 tmp_analyser = func(analysis_context.aminer_config, anomaly_event_handlers, target_path_list=item['paths'],
                                     persistence_id=item['persistence_id'],
                                     window_size=item['window_size'], num_windows=item['num_windows'],
