+++ source/root/usr/lib/logdata-anomaly-miner/aminer/schemas/normalisation/AnalysisNormalisationSchema.yml
@@ -57,6 +57,9 @@
                 'min_allowed_time_diff': {'type': 'float', 'default': 5.0},
                 'lower_limit': {'type': ['integer', 'float']},
                 'upper_limit': {'type': ['integer', 'float']},
+                'idf': {'type': 'boolean', 'default': False},
+                'norm': {'type': 'boolean', 'default': False},
+                'update_at_retrain': {'type': 'boolean', 'default': False},
                 'bin_size': {'type': 'integer'},
                 'bin_count': {'type': 'integer'},
                 'outlier_bins_flag': {'type': 'boolean', 'default': False},
