+++ source/root/usr/lib/logdata-anomaly-miner/aminer/schemas/validation/AnalysisValidationSchema.yml
@@ -107,11 +107,14 @@
                 },
                 {
                     'id': {'type': 'string', 'nullable': True, 'empty': False},
-                    'type': {'type': 'string', 'allowed': ['EventCountDetector'], 'required': True},
+                    'type': {'type': 'string', 'allowed': ['EventCountClusterDetector'], 'required': True},
                     'paths': {'type': 'list', 'schema': {'type': 'string', 'empty': False}, 'nullable': True},
                     'window_size': {'type': ['integer', 'float'], 'min': 0.001},
                     'num_windows': {'type': 'integer'},
                     'confidence_factor': {'type': 'float', 'min': 0, 'max': 1},
+                    'idf': {'type': 'boolean'},
+                    'norm': {'type': 'boolean'},
+                    'update_at_retrain': {'type': 'boolean'},
                     'persistence_id': {'type': 'string', 'empty': False},
                     'learn_mode': {'type': 'boolean'},
                     'output_logline': {'type': 'boolean'},
